fix: event delegation to fix duplicated api requests
