+++ static/js/settings.js
@@ -9,6 +9,7 @@ import {
 import { applyDarkMode, showSection, lastVisibleSectionId, openThemeCreator } from './ui.js'; // Need applyDarkMode, showSection, and lastVisibleSectionId
 import { initThemeCreator } from './theming.js';
 import { initializeTomSelect } from './ui-helpers.js';
+import { loadCustomInstructionsData, initializeCustomInstructions } from './custom-instructions.js';
 
 // --- State Variables ---
 export let currentSettings = { // Store loaded settings
@@ -351,6 +352,8 @@ function renderContainerHTML(container) {
 `;
     // Attach all event handlers and initialize components for this container
     attachContainerEventHandlers(container);
+    // Initialize the delegated event listeners for custom instructions ONCE
+    initializeCustomInstructions();
 }
 
 function attachContainerEventHandlers(container) {
@@ -802,361 +805,4 @@ export async function loadOllamaModels() {
     // This function is now a no-op because initializeSettings handles it.
 }
 
-// --- Custom Instructions Logic ---
-
-async function loadCustomInstructionsData(container) {
-    try {
-        const [instructions, sets] = await Promise.all([
-            apiRequest('/api/custom-instructions'),
-            apiRequest('/api/instruction-sets')
-        ]);
-        renderCustomInstructionsUI(instructions, sets, container);
-    } catch (error) {
-        console.error("Error loading custom instructions data:", error);
-        const instructionsContainer = container.querySelector('#instructions-list-container');
-        if (instructionsContainer) {
-            instructionsContainer.innerHTML = `<p class="error-message">Error loading instructions: ${error.message}</p>`;
-        }
-    }
-}
-
-function renderCustomInstructionsUI(instructions, sets, container) {
-    const instructionsContainer = container.querySelector('#instructions-list-container');
-    const setsContainer = container.querySelector('#instruction-sets-list-container');
-
-    if (!instructionsContainer || !setsContainer) return;
-
-    instructionsContainer.innerHTML = '';
-    instructions.forEach(inst => {
-        const instEl = document.createElement('div');
-        instEl.className = 'instruction-item';
-        instEl.dataset.instructionId = inst.id;
-
-        const subforumPills = inst.subforum_defaults.map(sf => `
-            <span class="subforum-pill" data-subforum-id="${sf.subforum_id}">
-                ${sf.name}
-                <button class="delete-pill-btn" data-instruction-id="${inst.id}" data-subforum-id="${sf.subforum_id}">&times;</button>
-            </span>
-        `).join('');
-
-        const hasSubforumDefaults = inst.subforum_defaults.length > 0;
-
-        instEl.innerHTML = `
-            <details>
-                <summary class="instruction-header">
-                    <strong class="instruction-name">${inst.name}</strong>
-                    <div class="instruction-controls">
-                        <button class="button-secondary button-small edit-instruction-btn">Edit</button>
-                        <button class="button-danger button-small delete-instruction-btn">Delete</button>
-                    </div>
-                </summary>
-                <div class="instruction-body">
-                    <p class="prompt-preview"><em>Prompt:</em> ${inst.prompt_text.substring(0, 150)}${inst.prompt_text.length > 150 ? '...' : ''}</p>
-                    <div class="setting-item">
-                        <label>Priority:</label>
-                        <input type="number" class="priority-input" value="${inst.priority}" style="width: 60px;">
-                    </div>
-                    <div class="setting-item">
-                        <label>Global Default:</label>
-                        <input type="checkbox" class="global-default-checkbox" ${inst.is_global_default ? 'checked' : ''}>
-                    </div>
-                    <div class="setting-item">
-                        <label>Subforum Default:</label>
-                        <input type="checkbox" class="subforum-default-checkbox" ${hasSubforumDefaults ? 'checked' : ''}>
-                    </div>
-                    <div class="subforum-input-container" style="display: ${hasSubforumDefaults ? 'block' : 'none'};">
-                        <div class="subforum-pills-container">${subforumPills}</div>
-                        <select class="subforum-search-input" placeholder="Add subforum..."></select>
-                    </div>
-                </div>
-            </details>
-        `;
-        instructionsContainer.appendChild(instEl);
-    });
-
-    // --- Instruction Sets UI --- 
-    setsContainer.innerHTML = '';
-    sets.forEach(set => {
-        const setEl = document.createElement('div');
-        setEl.className = 'instruction-set-item';
-        setEl.dataset.setId = set.id;
-
-        const instructionPills = set.instruction_ids.map(instId => {
-            const instruction = instructions.find(i => i.id === instId);
-            return `<span class="instruction-pill">${instruction ? instruction.name : 'Unknown'}</span>`;
-        }).join('');
-
-        setEl.innerHTML = `
-            <div class="instruction-set-header">
-                <strong>${set.name}</strong>
-                <div class="instruction-set-controls">
-                    <button class="button-secondary button-small edit-instruction-set-btn">Edit</button>
-                    <button class="button-danger button-small delete-instruction-set-btn">Delete</button>
-                </div>
-            </div>
-            <div class="instruction-set-body">${instructionPills}</div>
-        `;
-        setsContainer.appendChild(setEl);
-    });
-
-    attachCustomInstructionsEventHandlers(container, instructions, sets);
-}
-
-function attachCustomInstructionsEventHandlers(container, instructions, sets) {
-    const instructionsContainer = container.querySelector('#instructions-list-container');
-    const addInstructionBtn = container.querySelector('#add-instruction-btn');
-    const instructionModal = document.getElementById('instruction-modal');
-    const instructionForm = document.getElementById('instruction-form');
-    const instructionModalTitle = document.getElementById('instruction-modal-title');
-    const instructionIdInput = document.getElementById('instruction-id-input');
-    const instructionNameInput = document.getElementById('instruction-name-input');
-    const instructionPromptInput = document.getElementById('instruction-prompt-input');
-    const closeBtn = instructionModal.querySelector('.close-btn');
-
-    if (!instructionsContainer) return;
-
-    const openInstructionModal = (instruction = null) => {
-        if (instruction) {
-            instructionModalTitle.textContent = 'Edit Instruction';
-            instructionIdInput.value = instruction.id;
-            instructionNameInput.value = instruction.name;
-            instructionPromptInput.value = instruction.prompt_text;
-        } else {
-            instructionModalTitle.textContent = 'Add Instruction';
-            instructionForm.reset();
-            instructionIdInput.value = '';
-        }
-        instructionModal.style.display = 'block';
-    };
-
-    const closeInstructionModal = () => {
-        instructionModal.style.display = 'none';
-    };
-
-    addInstructionBtn.addEventListener('click', () => openInstructionModal());
-    closeBtn.addEventListener('click', closeInstructionModal);
-    window.addEventListener('click', (e) => {
-        if (e.target === instructionModal) {
-            closeInstructionModal();
-        }
-    });
-
-    instructionForm.addEventListener('submit', async (e) => {
-        e.preventDefault();
-        const id = instructionIdInput.value;
-        const data = {
-            name: instructionNameInput.value,
-            prompt_text: instructionPromptInput.value,
-            priority: 0, // Default value, will be overwritten if editing
-            is_global_default: false // Default value, will be overwritten if editing
-        };
-
-        const method = id ? 'PUT' : 'POST';
-        const url = id ? `/api/custom-instructions/${id}` : '/api/custom-instructions';
-
-        if (id) {
-            const originalInstruction = instructions.find(i => i.id == id);
-            if (originalInstruction) {
-                data.priority = originalInstruction.priority;
-                data.is_global_default = originalInstruction.is_global_default;
-            }
-        }
-
-        try {
-            await apiRequest(url, method, data);
-            closeInstructionModal();
-            loadCustomInstructionsData(container);
-        } catch (error) {
-            if (error.message.includes('already exists')) {
-                alert(`Error: ${error.message}`);
-            } else {
-                alert(`An unexpected error occurred: ${error.message}`);
-            }
-        }
-    });
-
-    instructionsContainer.addEventListener('click', async (e) => {
-        const instItem = e.target.closest('.instruction-item');
-        if (!instItem) return;
-        const instructionId = instItem.dataset.instructionId;
-
-        if (e.target.classList.contains('delete-pill-btn')) {
-            const subforumId = e.target.dataset.subforumId;
-            if (confirm(`Are you sure you want to remove this subforum default?`)) {
-                try {
-                    await apiRequest(`/api/custom-instructions/${instructionId}/subforum-defaults/${subforumId}`, 'DELETE');
-                    loadCustomInstructionsData(container);
-                } catch (error) {
-                    alert(`Error removing default: ${error.message}`);
-                }
-            }
-        } else if (e.target.classList.contains('delete-instruction-btn')) {
-            if (confirm('Are you sure you want to delete this instruction?')) {
-                try {
-                    await apiRequest(`/api/custom-instructions/${instructionId}`, 'DELETE');
-                    loadCustomInstructionsData(container);
-                } catch (error) {
-                    alert(`Error deleting instruction: ${error.message}`);
-                }
-            }
-        } else if (e.target.classList.contains('edit-instruction-btn')) {
-            const instruction = instructions.find(i => i.id == instructionId);
-            if (instruction) {
-                openInstructionModal(instruction);
-            }
-        }
-    });
-
-    instructionsContainer.addEventListener('change', async (e) => {
-        const instItem = e.target.closest('.instruction-item');
-        if (!instItem) return;
-        const instructionId = instItem.dataset.instructionId;
-        const instruction = instructions.find(i => i.id == instructionId);
-        if (!instruction) return;
-
-        if (e.target.classList.contains('priority-input') || e.target.classList.contains('global-default-checkbox')) {
-            const priority = instItem.querySelector('.priority-input').value;
-            const is_global_default = instItem.querySelector('.global-default-checkbox').checked;
-            
-            try {
-                await apiRequest(`/api/custom-instructions/${instructionId}`, 'PUT', {
-                    name: instruction.name,
-                    prompt_text: instruction.prompt_text,
-                    priority: parseInt(priority, 10),
-                    is_global_default: is_global_default
-                });
-                instruction.priority = parseInt(priority, 10);
-                instruction.is_global_default = is_global_default;
-            } catch (error) {
-                alert(`Error updating instruction: ${error.message}`);
-                loadCustomInstructionsData(container);
-            }
-        } else if (e.target.classList.contains('subforum-default-checkbox')) {
-            const inputContainer = instItem.querySelector('.subforum-input-container');
-            inputContainer.style.display = e.target.checked ? 'block' : 'none';
-        }
-    });
-
-    // --- Instruction Set Event Handlers ---
-    const setsContainer = container.querySelector('#instruction-sets-list-container');
-    const addSetBtn = container.querySelector('#add-instruction-set-btn');
-    const setModal = document.getElementById('instruction-set-modal');
-    const setForm = document.getElementById('instruction-set-form');
-    const setModalTitle = document.getElementById('instruction-set-modal-title');
-    const setIdInput = document.getElementById('instruction-set-id-input');
-    const setNameInput = document.getElementById('instruction-set-name-input');
-    const setInstructionsSelect = document.getElementById('instruction-set-instructions-select');
-    const setCloseBtn = setModal.querySelector('.close-btn');
-
-    const instructionOptions = instructions.map(inst => ({ value: inst.id, text: inst.name }));
-    const setInstructionsTomSelect = initializeTomSelect(setInstructionsSelect, {
-        options: instructionOptions,
-        plugins: ['remove_button'],
-    });
-
-    const openSetModal = (set = null) => {
-        setInstructionsTomSelect.clear();
-        setInstructionsTomSelect.clearOptions();
-        setInstructionsTomSelect.addOption(instructionOptions);
-
-        if (set) {
-            setModalTitle.textContent = 'Edit Instruction Set';
-            setIdInput.value = set.id;
-            setNameInput.value = set.name;
-            setInstructionsTomSelect.setValue(set.instruction_ids);
-        } else {
-            setModalTitle.textContent = 'Add Instruction Set';
-            setForm.reset();
-            setIdInput.value = '';
-        }
-        setModal.style.display = 'block';
-    };
-
-    const closeSetModal = () => {
-        setModal.style.display = 'none';
-    };
-
-    addSetBtn.addEventListener('click', () => openSetModal());
-    setCloseBtn.addEventListener('click', closeSetModal);
-    window.addEventListener('click', (e) => {
-        if (e.target === setModal) closeSetModal();
-    });
-
-    setForm.addEventListener('submit', async (e) => {
-        e.preventDefault();
-        const id = setIdInput.value;
-        const data = {
-            name: setNameInput.value,
-            instruction_ids: setInstructionsTomSelect.getValue()
-        };
-        const method = id ? 'PUT' : 'POST';
-        const url = id ? `/api/instruction-sets/${id}` : '/api/instruction-sets';
-
-        try {
-            await apiRequest(url, method, data);
-            closeSetModal();
-            loadCustomInstructionsData(container);
-        } catch (error) {
-            if (error.message.includes('already exists')) {
-                alert(`Error: ${error.message}`);
-            } else {
-                alert(`An unexpected error occurred: ${error.message}`);
-            }
-        }
-    });
-
-    setsContainer.addEventListener('click', async (e) => {
-        const setItem = e.target.closest('.instruction-set-item');
-        if (!setItem) return;
-        const setId = setItem.dataset.setId;
-
-        if (e.target.classList.contains('edit-instruction-set-btn')) {
-            const set = sets.find(s => s.id == setId);
-            if (set) openSetModal(set);
-        } else if (e.target.classList.contains('delete-instruction-set-btn')) {
-            if (confirm('Are you sure you want to delete this set?')) {
-                try {
-                    await apiRequest(`/api/instruction-sets/${setId}`, 'DELETE');
-                    loadCustomInstructionsData(container);
-                } catch (error) {
-                    alert(`Error deleting set: ${error.message}`);
-                }
-            }
-        }
-    });
-    container.querySelectorAll('.subforum-search-input').forEach(select => {
-        if (select.tomselect) return;
-
-        const instItem = select.closest('.instruction-item');
-        const instructionId = instItem.dataset.instructionId;
-
-        initializeTomSelect(select, {
-            valueField: 'subforum_id',
-            labelField: 'name',
-            searchField: 'name',
-            create: false,
-            load: async (query, callback) => {
-                try {
-                    const data = await apiRequest(`/api/subforums/search?q=${encodeURIComponent(query)}`);
-                    callback(data);
-                } catch (error) {
-                    console.error("Error fetching subforums:", error);
-                    callback([]);
-                }
-            },
-            onChange: async (value) => {
-                if (!value) return;
-                const subforumId = value;
-                try {
-                    await apiRequest(`/api/custom-instructions/${instructionId}/subforum-defaults`, 'POST', { subforum_id: subforumId });
-                    loadCustomInstructionsData(container);
-                } catch (error) {
-                    alert(`Error adding subforum default: ${error.message}`);
-                }
-                const ts = select.tomselect;
-                ts.clear();
-                ts.blur();
-            }
-        });
-    });
-}
+// All custom instructions logic has been moved to static/js/custom-instructions.js
